[CHANGE]Try async way

+++ Assets/Scripts/UnitScript.cs
@@ -5,8 +5,8 @@ public class UnitScript : MonoBehaviour
     [SerializeField]
     private Unit unit;
 
-    public int health;
-    public int attackDamage;
+    public float health;
+    public float attackDamage;
     public float attackSpeed;
     
     // Start is called once before the first execution of Update after the MonoBehaviour is created
